Initialization fix and cleanup
This commit fixed various issues revolving around initialization of
data that the new features in sealed secrets depends on. These
changes enable sealed secrets to survive pod death with the
keylist and blacklist in tact, as well as correctly copying all
current keys into the containers local key registry.
